requirements.txt
langchain
lanchain-openai
langchain-community
pydantic
wikipedia
python-dotenv
